backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
revyl[sdk]
python-dotenv
